+++ pool/certora_build.py
@@ -13,7 +13,7 @@ COMMAND = "just build"
 
 # JSON FIELDS
 PROJECT_DIR = (SCRIPT_DIR / "../").resolve()
-SOURCES = ["pool/src/pool/*.rs", "Cargo.toml"]
+SOURCES = ["pool/src/**/*.rs", "Cargo.toml", "pool/Cargo.toml"]
 EXECUTABLES = "target/wasm32-unknown-unknown/release/pool.wasm"
 
 VERBOSE = False
