+++ pool/confs/pool.conf
@@ -2,6 +2,7 @@
     "build_script": "../certora_build.py",
     "optimistic_loop": true,
     "process": "emv",
+    "loop_iter": "2",
     "precise_bitwise_ops": true,
     "rule": [
         "user_health",
