+++ pool/src/spec/mod.rs
@@ -1,4 +1,4 @@
 pub(crate) mod summaries;
-pub(crate) mod spec;
+pub(crate) mod rules;
 pub(crate) mod token;
 pub(crate) mod model;
